day 03 p 1
I probably wrote a bit too much, and also beginning to reach the limits of sane excel as I try to work around it's idiosyncrasies.

It would definitely be an interesting (🤔) challenge to reduce these formulas down to a single lambda that takes a range containing the input, but I guess I'll save that for next year.